Added option to export original images, instead of edited ones.
Exporting original images could be desired in a lot of situations, but
this addition came out of realizing this: iPhone photos are taken
upside down, with an orientation flag set in the exif data to allow
them to be displayed in the proper orientation. That's true for a lot
of cameras, not only the iPhone. iPhoto 9.4.3 actually rotates and
recompresses these images, and stores them in the iPhoto library under
ImagePath as an edited image. This is true even if you yourself hasn't
done any editing on the photo.
That means that using the script without setting the -o option, exports
the recompressed version from ImagePath.